+++ libMx/d_auto_filter.c
@@ -433,7 +433,7 @@ mxd_auto_filter_change_control( MX_AUTOSCALE *autoscale )
 	old_filter_setting = auto_filter->present_filter_setting;
 
 	MX_DEBUG( 2,("%s: change_control = %d",
-			fname, autoscale->change_control));
+			fname, (int) autoscale->change_control));
 
 	switch( autoscale->change_control ) {
 	case MXF_AUTO_NO_CHANGE:
@@ -467,7 +467,7 @@ mxd_auto_filter_change_control( MX_AUTOSCALE *autoscale )
 	default:
 		return mx_error( MXE_ILLEGAL_ARGUMENT, fname,
 			"Illegal value %d for change_control.",
-			autoscale->change_control );
+			(int) autoscale->change_control );
 	}
 
 	MX_DEBUG( 2,("%s: Changing filter '%s' from %#lx to %#lx.",
@@ -541,7 +541,7 @@ mxd_auto_filter_get_parameter( MX_AUTOSCALE *autoscale )
 		fname, autoscale->record->name,
 		mx_get_field_label_string( autoscale->record,
 			autoscale->parameter_type ),
-		autoscale->parameter_type));
+		(int) autoscale->parameter_type));
 
 	switch( autoscale->parameter_type ) {
 	default:
@@ -571,7 +571,7 @@ mxd_auto_filter_set_parameter( MX_AUTOSCALE *autoscale )
 		fname, autoscale->record->name,
 		mx_get_field_label_string( autoscale->record,
 			autoscale->parameter_type ),
-		autoscale->parameter_type));
+		(int) autoscale->parameter_type));
 
 	switch( autoscale->parameter_type ) {
 	default:
