+++ libMx/d_bluice_motor.c
@@ -585,7 +585,7 @@ mxd_bluice_motor_get_parameter( MX_MOTOR *motor )
 		fname, motor->record->name,
 		mx_get_field_label_string( motor->record,
 			motor->parameter_type ),
-		motor->parameter_type));
+		(int) motor->parameter_type));
 
 
 	switch( motor->parameter_type ) {
@@ -628,7 +628,7 @@ mxd_bluice_motor_set_parameter( MX_MOTOR *motor )
 		fname, motor->record->name,
 		mx_get_field_label_string( motor->record,
 			motor->parameter_type ),
-		motor->parameter_type));
+		(int) motor->parameter_type));
 
 	if ( foreign_motor->u.motor.is_pseudo ) {
 		return mx_error( MXE_UNSUPPORTED, fname,
